% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabtab.R
\name{launch}
\alias{launch}
\title{Launch dabtab apps}
\usage{
launch(package = "dabtab.basic", run = "viewer", state, ...)
}
\arguments{
\item{package}{dabtab package to start. One of "dabtab.basic", "dabtab.design", "dabtab.basics", "dabtab.model", "dabtab.multivariate", or "dabtab"}

\item{run}{Run a dabtab app in an external browser ("browser"), an Rstudio window ("window"), or in the Rstudio viewer ("viewer")}

\item{state}{Path to statefile to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Launch dabtab apps
}
\details{
See \url{https://dabtab-rstats.github.io/docs/} for dabtab documentation and tutorials
}
\examples{
\dontrun{
launch()
launch(run = "viewer")
launch(run = "window")
launch(run = "browser")
}

}
