% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_modifier.R
\name{my_theme_element}
\alias{my_theme_element}
\title{Generate a custom ggplot2 theme element.}
\usage{
my_theme_element(
  title_size,
  title_color,
  fill_color,
  axis_size,
  axis_color,
  font_family,
  label = "Input text: "
)
}
\arguments{
\item{title_size}{Font size for the title.}

\item{title_color}{Color for the title.}

\item{fill_color}{Background color.}

\item{axis_size}{Font size for axis labels.}

\item{axis_color}{Color for axis labels.}

\item{font_family}{Font family.}

\item{label}{Input text label (optional).}
}
\value{
A ggplot2 theme element.
}
\description{
Generate a custom ggplot2 theme element.
}
