% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabtab.R
\name{z_outlier}
\alias{z_outlier}
\title{Generate code to read a file}
\usage{
z_outlier(df, var, z_thresh)
}
\arguments{
\item{df}{Dataframe}

\item{var}{Column}

\item{z_thresh}{Z thresh value}
}
\description{
Generate code to read a file
}
\details{
Return two datasets before and after removing outliners
}
\examples{
  z_outliner(df, var="column A", z_thresh=2)
}
